Various changes to make v1.11.1 build properly.

+++ source/common/common/posix/thread_impl.cc
@@ -18,7 +18,7 @@ int64_t getCurrentThreadId() {
   pthread_threadid_np(nullptr, &tid);
   return tid;
 #elif defined(__sun)
-  return static_cast<int32_t>(pthread_self());
+  return (unsigned long)pthread_self();
 #else
 #error "Enable and test pthread id retrieval code for you arch in pthread/thread_impl.cc"
 #endif
